[BE] refactor: DTO 구조 리팩토링
- 연관있는 DTO 끼리 파일 구조화

- issue (#191)
