+++ server/src/domains/mentoring/dto/MentoringRequestDto.ts
@@ -1,0 +1,8 @@
+import { IsNumber } from 'class-validator';
+
+export class MentoringRequestDto {
+  @IsNumber()
+  mentorId: number;
+  @IsNumber()
+  groupId: number;
+}
