+++ server/src/domains/mentoring/dto/MentoringRequestListDto.ts
@@ -20,10 +20,3 @@ export class MentoringRequestParam {
   @IsNumber()
   mid: number;
 }
-
-export class MentoringCancelParam {
-  @IsNumber()
-  mentor: number;
-  @IsNumber()
-  group: number;
-}
